JavaScript calls for conditionalJS

The list of JavaScript calls that can be used as a `jsCall` argument of conditionalJS. All the actions are reversible. E.g. when using `disable` call and conditionalJS condition is false the opposite action to disable is called (removing disable attribute).

attachClass(class, when = TRUE)

disable(when = TRUE)

show(when = TRUE)

css(..., important = FALSE, when = TRUE)

animateVisibility(
  effectShow = "fadeIn",
  effectHide = "fadeOut",
  delay = 0,
  duration = 500,
  ignoreInit = TRUE,
  when = TRUE
)

custom(true = NULL, false = NULL)

Arguments

class

A css to be attached to (or detached from) the UI element.

when

Should the (primary) action be executed when `condition` is TRUE (when = TRUE, default) or FALSE (when = FALSE).

...

Named style properties, where the name is the property name and the argument is the property value. See css for more details.

important

Should `!important` rule be attached to the added css?

effectShow, effectHide

Animation effects used for showing and hiding element. Check .cssEffects object for possible options.

delay

Delay of animation start (in milliseconds).

duration

Duration of animation (in milliseconds).

ignoreInit

Should the animation be skipped when application is in initial state?

true, false

JS callback that should be executed when condition is true or false. Can be custom JS (wrapped into JS) or one of the custom-callbacks.

Details

The currently offered actions:

attachClass Add provided class to the UI element.

disable Add disable attribute to the UI element - usually results with disabling the input controller.

show Show/hide an element with a help of `visibility:hidden` rule. Comparing to conditionalPanel (which uses display:none) results with rendering an output even if hidden.

css Add css (inline) rule to the UI object. When condition is false, the rule is removed.

animateVisibility Show/hide an element in an animated way.

custom Define custom true and false callback.