Text area component

Contrary to textAreaInput the component is not a binding itself (doesn't send input to the server). Thanks to that, the component can store much more text value without slowing down the application.

If you want to access the component value on request please use valueButton.

textArea(
  inputId,
  value,
  label,
  width = "100%",
  height = "200px",
  resize = "default",
  readonly = FALSE
)

updateTextArea(session, inputId, value = NULL)

Arguments

inputId

Id of component.

value

Initial text area value or value to be updated.

label

Text area label.

width

Width of input area.

height

Height of input area.

resize

Text are directions where input field can be resized. Possible options are "default", "both", "none", "vertical" and "horizontal".

readonly

If TRUE, providing custom values will be turned off.

session

Shiny session object.

Value

Nested list of `shiny.tag` objects defining html structure of the component, or no value in case of usage of `updateTextArea` method.